feat(uart): support 1.5 stop bits and update stop bit datatype
- Updated `lexerDefinition.l` and `parserDefinition.y` to support the UART stop bit value `1.5`.
- Changed the stop bits datatype from `uint8_t` to `float` in `astEnums.h`.
- Updated stop bit setter functions in `astBuild.c` and `astBuild.h` to use the new datatype.
- Updated `astPrint.c` to support printing the updated stop bit datatype.

+++ generator/astBuild.c
@@ -846,7 +846,7 @@ void ast_module_builder_set_uart_databits(int line_nr, ast_module_builder_t* mod
  * 
  * @note Logs an error if the stop bits have already been set, if the module kind is not UART, stopbits exceed uint8_t max, or if any parameter is NULL.
  */
-void ast_module_builder_set_uart_stopbits(int line_nr, ast_module_builder_t* module_builder, uint32_t stopbits){
+void ast_module_builder_set_uart_stopbits(int line_nr, ast_module_builder_t* module_builder, float stopbits){
   if(module_builder == NULL)
     log_error("ast_module_builder_set_uart_stopbits", 0, "AST module builder is NULL.");
