+++ generator/astBuild.h
@@ -44,7 +44,7 @@ void ast_module_builder_set_uart_tx_pin(  int line_nr, ast_module_builder_t* mod
 void ast_module_builder_set_uart_rx_pin(  int line_nr, ast_module_builder_t* module_builder, pin_t rx_pin);
 void ast_module_builder_set_uart_baudrate(int line_nr, ast_module_builder_t* module_builder, uint32_t baudrate);
 void ast_module_builder_set_uart_databits(int line_nr, ast_module_builder_t* module_builder, uint32_t databits);
-void ast_module_builder_set_uart_stopbits(int line_nr, ast_module_builder_t* module_builder, uint32_t stopbits);
+void ast_module_builder_set_uart_stopbits(int line_nr, ast_module_builder_t* module_builder, float stopbits);
 void ast_module_builder_set_uart_parity(  int line_nr, ast_module_builder_t* module_builder, uart_parity_t parity);
 
 #endif //__AST_BUILD_H__
